flytt logikk for automatisert utbetaling
